% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARP_test.R
\name{ARI}
\alias{ARI}
\title{Calculate the adjusted rand index}
\usage{
ARI(w, rowColor)
}
\arguments{
\item{w}{the ground-truth clusters}

\item{rowColor}{the predicted clusters}
}
\description{
This function is to calculate the performance of the algorithm by 5 metrics, including the Rand index, HA, MA, FM and Jaccard.
}
\examples{
finalmetrics = ARI(gtc, finalrowColor)

}
