% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARP_test.R
\name{SHARP_test}
\alias{SHARP_test}
\title{Run SHARP for single-cell RNA data clustering}
\usage{
SHARP_test(E, gtc, K, p)
}
\arguments{
\item{E}{input single-cell expression matrix}

\item{gtc}{the ground-truth clusters}

\item{K}{number of applications of random projection}

\item{p}{the dimension to be reduced to}
}
\description{
SHARP: \strong{S}ingle-cell RNA-Seq \strong{H}yper-fast and \strong{A}ccurate clustering via ensemble \strong{R}andom \strong{P}rojection.
}
\examples{
enresults = SHARP_test(E, gtc, K = 15)

}
